Link d3d8_tests to d3d8.lib
This is somewhat inconvenient. The Windows SDK does not ship d3d8.lib, so
we have to use ours. That means by default out d3d8.dll will also be there
and will have to be deleted manually every time the test is recompiled to
run the test with native d3d8.